addEvent & deleteEvent Update
An extra check is needed to see whether the event exists in the list or it null. We don't want duplicate records or empty records.
So, if the event already exists, a warning message appears, and it is not added to the list again.
A message also appears if the organizer wants to delete an event not included in the relative list.